Add modules, add function to display chars
